Enabling both fakeredis and real redis tests

+++ spec/spec_helper.rb
@@ -10,7 +10,12 @@ require 'rubygems'
 require 'bundler/setup' if File.exists?(ENV['BUNDLE_GEMFILE'])
 require 'pause'
 require 'pry'
-require 'pause/redis/adapter'
+
+if ENV['PAUSE_REAL_REDIS']
+  require 'pause/redis/adapter'
+else
+  require 'fakeredis/rspec'
+end
 
 RSpec.configure do |config|
   config.run_all_when_everything_filtered = true
@@ -21,7 +26,9 @@ RSpec.configure do |config|
   # the seed, which is printed after each run.
   #     --seed 1234
   config.order = 'random'
-  config.before(:example) do
-    Pause::Redis::Adapter.redis.flushdb
+  if ENV['PAUSE_REAL_REDIS']
+    config.before(:example) do
+      Pause::Redis::Adapter.redis.flushdb
+    end
   end
 end
